fix(apps): satisfy swift, periphery, and lint gates for native digests
